requirements.txt
aioimgur
asyncpg
line.py @ git+https://github.com/seriaati/line.py
nicegui
tortoise-orm
qrcode
Pillow